Added unit tests for AbstractPropertyConstraint
Improved the unit test coverage in the package
org.openecomp.sdc.be.model.tosca.constraints

Issue-ID: SDC-2327

+++ catalog-model/src/main/java/org/openecomp/sdc/be/model/tosca/ToscaType.java
@@ -163,7 +163,7 @@ public enum ToscaType {
 			case BOOLEAN:
 				return Boolean.valueOf(value);
 			case FLOAT:
-				return Double.valueOf(value);
+				return Float.valueOf(value);
 			case INTEGER:
 				return Long.valueOf(value);
 			case TIMESTAMP:
